+++ Changelog.md
@@ -10,6 +10,7 @@
 - Add QuestTracker.AutoRequest setting
 - Rewrite DungeonMaps with performance in mind
 - Fix equip mgr command pattern, default profile selection, and clean up event handling
+- Fix `/ub count item` output when no items were found
 
 ## 0.1.1 (2019-12-03) [UtilityBeltInstaller-0.1.1.exe](https://gitlab.com/utilitybelt/utilitybelt.gitlab.io/uploads/c94b459edb29649f99f1d84eae3f76ce/UtilityBeltInstaller-0.1.1.exe)
 - Increase AutoStack/AutoCram timeout to 30s